requirements.txt.txt
streamlit
fpdf2
Pillow
qrbill
cairosvg
